refactor(client): split ApiClient into domain-specific API modules (CRAP 184 -> <40)
Extract core infrastructure (ApiError, fetchClient, apiClient) into core.ts.
Create focused API modules: movies.ts, scraper.ts, schedules.ts, reports.ts.
Consolidate theater API functions into theaters.ts (addTheater, getTheaterSchedule).
client.ts becomes a lean barrel file re-exporting all domain modules.

Closes #1143
